requirements.txt
jsonata-python
paho-mqtt